+++ VoxTerGenEngine/include/world/ChunkManager.hpp
@@ -10,12 +10,7 @@
 
 struct ivec2_hash 
 {
-    std::size_t operator()(const glm::ivec2& v) const noexcept 
-    {
-        const std::size_t h1 = std::hash<int>{}(v.x);
-        const std::size_t h2 = std::hash<int>{}(v.y);
-        return h1 ^ (h2 << 1);
-    }
+	std::size_t operator()(const glm::ivec2& vec) const noexcept;
 };
 
 class ChunkManager
@@ -26,13 +21,18 @@ private:
 public:
 	ChunkManager();
 
-	//Block& NeighborAt(glm::ivec2 chunk_coord, int x, int y, int z, Direction dir) const;
+    void InitChunks(int chunk_radius);
+    
+    void InitChunkBlocks(Chunk& chunk);
+
+	void BuildAllChunkMeshes();
+
+	Block WorldBlockQuery(glm::ivec2 current_chunk_coord, const glm::ivec3& block_coords) const;
 
 	const Chunk* GetChunkAt(glm::ivec2 chunk_coord) const;
 
 	std::unordered_map<glm::ivec2, std::unique_ptr<Chunk>, ivec2_hash>& Chunks() { return chunks_; }
 
-	//Chunk& ChunkAt(glm::ivec2 chunk_coord);
 };
 
 #endif // CHUNK_MANAGER_HPP