+++ VoxTerGenEngine/src/core/Settings.cpp
@@ -1,1 +1,7 @@
 #include "core/Settings.hpp"
+#include "utils/Constants.hpp"
+
+Settings::Settings() : chunk_radius_(constants::chunk::default_radius)
+{
+
+}